Untangle dependencies to fix tests
