+++ src/models/beacon.ts
@@ -17,7 +17,7 @@ limitations under the License.
 import { MBeaconEventContent } from "../@types/beacon";
 import { M_TIMESTAMP } from "../@types/location";
 import { BeaconInfoState, BeaconLocationState, parseBeaconContent, parseBeaconInfoContent } from "../content-helpers";
-import { MatrixEvent } from "../matrix";
+import { MatrixEvent } from "./event";
 import { sortEventsByLatestContentTimestamp } from "../utils";
 import { TypedEventEmitter } from "./typed-event-emitter";
